Unify detector interfaces and fix/pass tests
